backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.5
requests==2.31.0
aiohttp==3.9.1